+++ include/LIEF/Abstract/DebugInfo.hpp
@@ -15,7 +15,11 @@
 #ifndef LIEF_DEBUGINFO_H
 #define LIEF_DEBUGINFO_H
 #include <memory>
+#include <cstdint>
+#include <string>
+
 #include "LIEF/visibility.h"
+#include "LIEF/optional.hpp"
 namespace LIEF {
 
 class Binary;
@@ -24,6 +28,13 @@ namespace details {
 class DebugInfo;
 }
 
+/// This class provides a generic interface for accessing debug information
+/// from different formats such as DWARF and PDB.
+///
+/// Users can use this interface to access high-level debug features like
+/// resolving function addresses.
+///
+/// See: LIEF::pdb::DebugInfo, LIEF::dwarf::DebugInfo
 class LIEF_API DebugInfo {
   public:
   friend class Binary;
@@ -56,6 +67,9 @@ class LIEF_API DebugInfo {
     }
     return nullptr;
   }
+
+  /// Attempt to resolve the address of the function specified by `name`.
+  virtual optional<uint64_t> find_function_address(const std::string& name) const = 0;
 
   protected:
   std::unique_ptr<details::DebugInfo> impl_;